Reorganização dos diretórios para disponibilizar o pacote python.
