+++ .gitignore
@@ -7,6 +7,7 @@ venv/
 *.egg-info/
 *.pytest_cache/
 .coverage
+*build/
 
 # Ambiente
 .env
