--- src/gerador_dados/__init__.py
@@ -1,3 +1,0 @@
-from src.gerador_dados.gerador import Gerador
-
-__all__ = ["Gerador"]
