+++ src/sistema_antifraude/gerador_dados/__init__.py
@@ -1,0 +1,3 @@
+from sistema_antifraude.gerador_dados.gerador import Gerador
+
+__all__ = ["Gerador"]
