+++ src/sistema_antifraude/tests/test_gerador_dados.py
@@ -3,9 +3,9 @@ import uuid
 
 from faker import Faker
 
-from src.gerador_dados.componentes.dado_gerado import DadoGerado
-from src.gerador_dados.componentes.gerador_dados_legitimos import GeradorDadosLegitimos
-from src.gerador_dados.componentes.gerador_dados_fraudados import GeradorDadosFraudados
+from sistema_antifraude.gerador_dados.componentes.dado_gerado import DadoGerado
+from sistema_antifraude.gerador_dados.componentes.gerador_dados_legitimos import GeradorDadosLegitimos
+from sistema_antifraude.gerador_dados.componentes.gerador_dados_fraudados import GeradorDadosFraudados
 
 
 class TestDadoGerado:
